Simulated system include, and added macro test
